+++ crates/sncast/src/starknet_commands/account/add.rs
@@ -61,7 +61,7 @@ pub async fn add(
     rpc_url: &str,
     account: &str,
     accounts_file: &Utf8PathBuf,
-    path_to_scarb_toml: &Option<Utf8PathBuf>,
+    manifest_path: &Option<Utf8PathBuf>,
     package_name: &Option<String>,
     provider: &JsonRpcClient<HttpTransport>,
     add: &Add,
@@ -94,7 +94,7 @@ pub async fn add(
     write_account_to_accounts_file(account, accounts_file, chain_id, account_json.clone())?;
 
     if add.add_profile {
-        let manifest_path = ensure_scarb_manifest_path(path_to_scarb_toml)?;
+        let manifest_path = ensure_scarb_manifest_path(manifest_path)?;
         let config = CastConfig {
             rpc_url: rpc_url.into(),
             account: account.into(),